Add files via upload
The index page leads to a portal that leads to assignments due throughout the course of the class. Assignment M2 was to create a sample of different ways to apply CSS to various elements. Assignment M2.2 was to create a form that would send an email once you hit submit.